# portable_env/python-portable/workspace/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FNO2 Model API",
    "description": "FastAPI-based web service for FNO2 velocity field prediction",
    "version": "1.0.0"
  },
  "paths": {
...
    "/predict": {
      "post": {
        "summary": "Predict Velocity Field",
        "operationId": "predict_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reload": {
      "post": {
        "summary": "Reload Model",
        "operationId": "reload_model_reload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReloadRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReloadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PredictRequest": {
        "properties": {
          "input": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "items": {
                      "type": "number"
                    },
                    "type": "array"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input",
            "description": "Combined input data"
          },
          "u": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "U",
            "description": "U velocity data"
          },
          "v": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "V",
            "description": "V velocity data"
          },
          "height": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height",
            "description": "Cavity height (m)"
          },
          "width": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width",
            "description": "Cavity width (m)"
          },
          "vel_top": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vel Top",
            "description": "Top wall velocity (m/s)"
          },
          "density": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Density",
            "description": "Fluid density (kg/m3)"
          },
          "viscosity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viscosity",
            "description": "Dynamic viscosity (Pa*s)"
          },
          "dx": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dx",
            "description": "Grid spacing in x-direction (m)"
          },
          "dy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dy",
            "description": "Grid spacing in y-direction (m)"
          },
          "denormalize": {
            "type": "boolean",
            "title": "Denormalize",
            "description": "Whether to denormalize the output",
            "default": true
          },
          "return_full": {
            "type": "boolean",
            "title": "Return Full",
            "description": "Whether to return full raw output",
            "default": false
          }
        },
        "type": "object",
        "title": "PredictRequest"
      },
      "PredictResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the prediction was successful",
            "default": true
          },
          "inference_time_ms": {
            "type": "number",
            "title": "Inference Time Ms",
            "description": "Inference time in milliseconds"
          },
          "output_shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Output Shape",
            "description": "Shape of the output tensor"
          },
          "u": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "U",
            "description": "Predicted U velocity"
          },
          "v": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "V",
            "description": "Predicted V velocity"
          },
          "raw_output": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "items": {
                      "type": "number"
                    },
                    "type": "array"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Output",
            "description": "Raw model output"
          }
        },
        "type": "object",
        "required": [
          "inference_time_ms",
          "output_shape"
        ],
        "title": "PredictResponse"
      },
      "ReloadRequest": {
        "properties": {
          "model_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Path",
            "description": "New model path"
          },
          "framework": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Framework",
            "description": "Framework: auto, pytorch, or paddle"
          }
        },
        "type": "object",
        "title": "ReloadRequest"
      },
      "ReloadResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the reload was successful"
          },
          "model_path": {
            "type": "string",
            "title": "Model Path",
            "description": "New model path"
          },
          "framework": {
            "type": "string",
            "title": "Framework",
            "description": "Framework used"
          },
          "device": {
            "type": "string",
            "title": "Device",
            "description": "Device used"
          },
          "model_info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Model Info",
            "description": "Detailed model information"
          },
          "normalization_params": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Normalization Params",
            "description": "Normalization parameters"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Current timestamp"
          }
        },
        "type": "object",
        "required": [
          "success",
          "model_path",
          "framework",
          "device",
          "model_info",
          "normalization_params",
          "timestamp"
        ],
        "title": "ReloadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}